Backend_django/requirements.txt
Django~=5.0.4
python-dotenv~=1.0.1
djangorestframework~=3.15.1
pillow~=10.3.0